+++ R/ifcb_get_shark_example.R
@@ -1,7 +1,7 @@
 #' Get Shark Column Example
 #'
 #' This function reads a SHARK submission example from a file included in the package.
-#' This format is used for submitting IFCB data to \url{https://shark.smhi.se/}.
+#' This format is used for submitting IFCB data to \url{https://shark.smhi.se/en/}.
 #'
 #' @return A data frame containing example data following the SHARK submission format.
 #' @export